修正社区的comment_num和view_num字段返回字符串类型数据导致的解析错误问题

+++ ExilumBBS/Models/DTO/PostPreview.cs
@@ -21,7 +21,7 @@ namespace ExilumBBS.Models.DTO
 
         [JsonPropertyName("comment_num")]
 
-        public int CommentNum { get; set; }
+        public long CommentNum { get; set; }
 
 
         [JsonPropertyName("content")]
